requirements.txt
Flask==2.1.3
tensorflow==2.5.0
numpy==1.19.5
opencv-python==4.5.3.56
gunicorn
Werkzeug==2.1.2
Pillow
scipy